Improvements to macros.
For now, we are going to disallow macro reference clicks which do not uniquely determine an orientation.  We might not be so Draconian in the future (see issue 23).

This change also marks that we are using grips in the macro files.  This is so we won't have to rev the file format later when/if we choose to support sticker reference clicks for macros (see issue 24).