Show total num paints in upper-right when in debug mode.
When in debug mode, e.g. "(123 paints)" is shown
in the upper-right corner of the picture.

In my perception, this gives instant clarity on a lot of repaint
and speed issues which were previously elusive.
E.g. it's now very easy to see the effect of keeping or removing
the semicolon that is mentioned in issue #136.