Fix single out-of-order polygon drawing screwup.
One arbitrary polygon was inappropriately getting drawn after (in front of)
everything else.  I accidentally introduced this bug when I upgraded
to latest donhatchsw libraries for the new polytope engine
(4d6359971d5df93da60c8553bd88e5c53cd96e5d).

Details:

The original code was using the following old
shouldn't-have-ever-been-public function:
  com.donhatchsw.util.SortStuff.sortRange(Object[] list, int firstindex, int lastindex, Comparator comparator)
The new code uses this instead:
  com.donhatchsw.util.SortStuff.sort(Object[] list, int i0, int n, Comparator comparator)

Note that the indexing convention is different, which I didn't notice :-(
In fact this accident-waiting-to-happen was called out in the old code function's comment:
  //
  // The actual recursive algorithm...
  // XXX actually it's messed up that this is public, because the interpretation of j is surprising-- it's the last index rather than the length
  //
  public static void sortRange(Object list[], int i, int j, Comparator comparator)

So the caller needed to change from this:
  com.donhatchsw.util.SortStuff.sortRange(drawList, 0, drawListSize - 1, ...);
to this:
  com.donhatchsw.util.SortStuff.sort(drawList, 0, drawListSize, ...);