Remove 1/2 pixel jump when antialiasing turns on/off.
This fixes issue #138.

Details:
Previously, antialiased polygon fill was offset 1/2 pixel
to the upper-left compared to the other 3 types of stuff
(antialiased outlines, non-antialiased fill, non-antialiased outlines).
This change turns on the voodoo STROKE_PURE hint for everything;
the effect is that those other 3 get adjusted,
so that now all 4 types of stuff match.
For a detailed study, see:
https://stackoverflow.com/questions/7701097/java-graphics-fillpolygon-how-to-also-render-right-and-bottom-edges/63645061#answer-63645061