Slight refactor of PuzzleDescription/PolytopePuzzleDescription.
Moved definition of ProgressCallbacks from concrete subclass PolytopePuzzleDescription
into interface PuzzleDescription.
In the upcoming new version of these classes, all classes implementing this interface
will be required to have the same constructor signature, so that they can be
uniformly serialized/deserialized;
so the types for that signature need to be defined in the base interface.