Introduce concept of doubleLength, independent of intLength.
Background:

PolytopePuzzleDescription will be upgraded soon to a new version
from donhatchsw, but that will be a significant chunk of work.
So, for short term gratification, I'm teaching the existing/old version
of this class the new better way to express puzzle lengths and cut depths,
as the new/future version of this class will think of these things.

This is throwaway work, in a sense, since this old version of
PolytopePuzzleDescription is going to be replaced.
But I'm doing it this way for two reasons:
(1) it will allow us to express a nicer selection of new puzzles,
sooner rather than later;
(2) it changes PolytopePuzzleDescription's API to be closer to what it will be,
getting that out of the way so that the rest of the port, later, will be a bit
less painful and disruptive for the callers.

Changes:

The new way of doing lengths and cut depths is as follows.
There is an intLength controlling the number of cuts, as always
(that is, number of cuts = intLength-1),
and a doubleLength, whose inverse determines the cut spacing.
In the prescription string, doubleLength may be given in parens;
e.g. "{4,3,3} 3(10.0)" means the standard hypercube of length 3,
but with very shallow cuts (1/10 of the way along an edge).
If doubleLength is omitted, it defaults to intLength;
so for example the standard puzzle can still be expressed as "{4,3,3} 3",
which can now be thought of as shorthand for "{4,3,3} 3(3.0)".

This can be particularly useful for triangles and tets, for which
3(4.0) or 3(5.0) is nicer in some ways than just 3.
So I added some of these doubleLength options to the menu, for existing shapes:
  "{3,3,3} 3(4.0)", "{3,3,3} 3(5.0)"
  "{3}x{3} 3(4.0)",
  "{3}x{4} 3(4.0)",
  "{3}x{5} 3(4.0)",
  "{3,3}x{} 3(4.0)" "{3,3}x{} 3(5.0)"
  "{4,3,3} 3(10.0)"  (hypercube with shallow cuts)
  "{3,4}x{} 3(4.0)"  (octahedral prism)
  "{3,4,3} 3(4.0)"  (24-cell)
As well as the following popular variation on cut depths for dodecahedral facets:
  "{5,3}x{} 3(2.5)"
  "{5,3,3} 3(2.5)"