Fix not-well-handled exception on bad length in "Invent my own" processing.
Previously, if the user gave a bad length in the "Invent my own"
dialog, it resulted in a non-graceful exception
(which, I'm guessing, most users won't even see?).
E.g. entering "{4,3,3} x" sent the following to my terminal:

 Exception in thread "AWT-EventQueue-0" java.lang.NumberFormatException: For input string: "x"
  at java.base/jdk.internal.math.FloatingDecimal.readJavaFormatString(FloatingDecimal.java:2054)
  at java.base/jdk.internal.math.FloatingDecimal.parseDouble(FloatingDecimal.java:110)
  at java.base/java.lang.Double.parseDouble(Double.java:543)
  at com.superliminal.magiccube4d.PuzzleManager.initPuzzle(PuzzleManager.java:108)
  ...

Now, it reports the error gracefully, same as when the user forgot the length string.