make topological fingerprints faster and better.
TL;DR:
Fingerprinting is no longer slow, and it now correctly recognizes that:
   "{5,3,3} 2" == "{5,3,3} 3"
   "{4,3,3} 2" != "{4,3,3} 3"
Probably still not 100% perfect, but I don't know of any cases that fail at the moment.

Details:

Previously the puzzle's topological fingerprint contained
the sliced polytope structure, including number of verts,edges,polys.
That was expensive to compute, and was wrong anyway,
since it incorrectly said "{5,3,3} 2" and "{5,3,3} 3" were different.

Now, instead of the full sliced polytope structure,
we use just the number of stickers and floor(intLength/2).
(And we still include the original polytope structure fingerprint.)

So the fingerprint now consists of:
   - fingerprint of original polytope
   - nStickers
   - "number of cuts" parallel to each face; that is, floor(intLength/2).