Puzzle fingerprints, attempt #3.  Maybe it's good now.
TL;DR:
The previous attempt at fingerprints failed to detect that:
  "{3}x{4} 3(3.25)" != "{3}x{4} 3(2.75)".
(Although these two puzzles look very similar at first glance,
the latter is a much harder puzzle, since the center
stickers of the tet-prism faces aren't stationary.)
This new version of fingerprinting is able to tell these two puzzles apart.

Caveat: fingerprinting works only if all edges in the polytope are of the same length
(which is true for all uniform polytopes, which includes products of uniforms).
Some cases where this property doesn't hold:
    "(1.0)3(0)3(2.0)*{} 3"
    "frucht*{} 3"
In such cases, the human readable fingerprint function
will return a string containing the substring "NOT FINGERPRINTABLE",
and the fingerprint digest function will return null.

DETAILS:

I added a "face type to slice sticker counts" section to the human-readable fingerprint,
showing how the stickers are distributed among the slices.

So the full human-readable fingerprints for the two puzzles
mentioned above now look like the following.
Notice the difference, at the very end of the two human-readable fingerprints,
showing that, when twisting a cubic face, the distribution of stickers
among the slices is different in the two puzzles; specifically,
in the second puzzle, twisting a cubic face moves more stuff.

"{3}x{4} 3(3.25)"
    Puzzle fingerprint human readable =
        original polytope:
            structure summary:
                12 verts
                24 edges
                19 polygons (15 quads, 4 triangles)
                7 3-cells (4*{6 verts, 9 edges, 5 polygons (3 quads, 2 triangles)}, 3*{8 verts, 12 edges, 6 quads})
            vertex figures and counts:
                "4 verts, 6 edges (1 3-chord, 5 4-chords), 4 polygons (2*{3 verts, 3 edges (1 3-chord, 2 4-chords)}, 2*{3 verts, 3 edges (3 4-chords)})": 12
        floor(intLength/2) = 1
        number of stickers: 165
        face type to slice sticker counts:
            "6 verts, 9 edges, 5 polygons (3 quads, 2 triangles)": 62,41,62
            "8 verts, 12 edges, 6 quads": 81,84
    Puzzle fingerprint digest = fa5b7f484aabc444d5638deff729b42e8c0d391c

"{3}x{4} 3(2.75)"
    Puzzle fingerprint human readable =
        original polytope:
            structure summary:
                12 verts
                24 edges
                19 polygons (15 quads, 4 triangles)
                7 3-cells (4*{6 verts, 9 edges, 5 polygons (3 quads, 2 triangles)}, 3*{8 verts, 12 edges, 6 quads})
            vertex figures and counts:
                "4 verts, 6 edges (1 3-chord, 5 4-chords), 4 polygons (2*{3 verts, 3 edges (1 3-chord, 2 4-chords)}, 2*{3 verts, 3 edges (3 4-chords)})": 12
        floor(intLength/2) = 1
        number of stickers: 165
        face type to slice sticker counts:
            "6 verts, 9 edges, 5 polygons (3 quads, 2 triangles)": 62,41,62
            "8 verts, 12 edges, 6 quads": 93,72
    Puzzle fingerprint digest = 6fc138d7ec3363035672d57ec120bac85bdbcf3a