Fix `dflt` param in 2nd PropRadioButton in each group.
I think this is right, for two reasons.

Reason #1: If I understand correctly,
calling PropertyManager.getBoolean(propname, true) says
"the default for this property is true",
in which case it doesn't make sense to then call
PropertyManager.getBoolean(propname, false) in the same program.

Reason #2: Temporarily commenting out the ButtonGroup stuff
revealed that the radiobuttons were given inconsistent initial values
(i.e. both ticked).  So it seems the ButtonGroup was masking this error.