Fix placement of checkboxes in `General` section part of prefs editor.
Now I think everything in the Preference editor looks good, except for the sliders.

Details:

A couple of components in this section were using a hack
to get left-justified: that is, make an intermediate JPanel
containing the component followed by a horizontal glue on the right.
I made a little function do that, and made all the checkboxes
and labels call it, so they are now all reasonably aligned:
that is, everything all the way to the left,
except for the radio buttons which have 30 pixels of indentation.