Add radiobuttons to control whether antialiasing is "when still" (default) or "always".
For some users (Roice) antialiasing seems to be free,
and there may soon be an improvement that will make it
very cheap for everyone, so it makes sense for the user
to be able to choose to antialias "always".

At some point in the future, it may make sense
to remove the "when still" option entirely
(i.e. the two choices will then be always-off or always-on),
but not yet.