listen to "the model" (underlying property) instead of "the view" (PropCheckBox).
Enabledness of "All Moves" and "Just Macros" radiobuttons
was bound to the checked-state of the "Quick Moves" checkbox.
Better to bind them to the underlying "quickmoves" property,
for correct response if the property value were, hypothetically,
changed in some other way.

Similary for enabledness of "when still" and "always" radiobuttons:
they now listen to the "antialias" property instead of to the checkbox
that happens to be bound to that property.

Also, make PropCheckbox and PropRadioButton stay in sync with the property they control
(i.e. change the binding from 1-way to 2-way), for similar reasons.
In particular, multiple checkboxes/radiobuttons bound to the same property will now stay in sync.
I didn't do the same for PropFloatSlider since that would be tricker and full of bugs.

Tested:
Temporarily added several extra copies of the "Allow Antiasing" and "Quick Moves" checkboxes.
They now stay in sync with each other, and checking any of them properly enables/disables
the corresponding two radiobuttons.