avoid spurious repaints of main view when hovering over Debugging checkbox.
Fixed a couple of things that were listening to "the view"
(Debugging checkbox) when they should have been listening
to "the model" (the underlying DEBUGGING property).

One of these was a change listener that was attached to the Debugging checkbox.
It was intended to go off only when the checkbox was toggled, but unfortunately
it was going off when simply hovering over the checkbox, too,
causing unnecessary sometimes-expensive repaints.

I noticed this because the frame counter was getting incremented for no apparent reason
when simply hovering over menus.