experimental: added a "Use back buffer" checkbox, to speed up antialiasing (!)
This experimental checkbox is visible only when in Debugging mode.

Details:

I noticed that antialiasing, which we always thought was super expensive,
becomes miraculously fast (on my Linux machine, anyway)
if we simply render to an auxiliary back buffer
instead of directly to the view Component.

To facilitate testing this out,
I've added a checkbox to the Preferences panel "Use back buffer (EXPERIMENTAL)",
visible only when in Debugging mode.
When this checkbox is ticked, rendering is done into a BufferedImage before being copied to the View component.
(This is in addition to whatever double-buffering swing is doing by default.)

Note that rendering into a BufferedImage can look subtlely different
from rendering directly into the component; for details, see
see https://stackoverflow.com/questions/7701097/java-graphics-fillpolygon-how-to-also-render-right-and-bottom-edges/63645061#answer-63645061 ,
and keep in mind that we have the STROKE_CONTROL hint set to STROKE_PURE
(since that seems to be the way to minimize jarring when toggling antialiasing).
In particular, try this combination:
  - "Allow Antialiasing" off
  - "Draw Outlines" on
  - toggle "Use back buffer" on and off;
    notice that, when it's on,
    the outlines shift 1/2 pixel down and to the right.
    (That's on my Linux machine.)